muevo eventos del temporizador al controlador
